Fix date marshalling/unmarshalling with offset
We have a bug since day one where we don't handle properly date with offset 🙈.

We have a custom fr.vidal.oss.jaxb.atom.core.DateAdapter to convert `String` to `Date` and vice versa.
The problem is that `SimpleDateFormat` used in the adapter was not handling properly the offset.
It was considaring everything as UTC where as the TimeZone used was the default system timezone.

So if we try to unmarshall `2023-10-10T10:10:10+02:00` we have 'null' date (because the formatter only understand 'Z' pattern 🤦‍♂️).
When we marshall a date, we were falsy returning 'Z' as offset whatever the system timezone was.

This commit fix the String parsing by taking into account the offset.
It fixes the marshalling by converting an instant to UTC (we have kept the fixed 'Z' offset for compatibility).

Fixes #172